# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.0.1",
    "title": "Winding tree API",
    "description": "Node js API for Winding Tree",
    "contact": {
      "name": "Winding tree",
      "url": "https://windingtree.com/"
    },
    "license": {
      "name": "Apache-2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    }
  },
...
  "schemes": [
    "http"
  ],
  "host": "localhost:3000",
  "basePath": "/",
...
  "paths": {
...
    "/hotels/{hotelAddress}/unitTypes": {
      "post": {
        "tags": [ "Unit types" ],
        "summary": "Add unit type to hotel",
        "description": "Add an unit type to the hotel.",
        "parameters": [
          {
            "in": "path",
            "name": "address",
            "type": "string",
            "required": true,
            "description": "Hotel address."
          },
          {
            "name": "params",
            "in": "body",
            "required": true,
            "schema": {
              "required": [
                "password",
                "type"
              ],
              "properties": {
                "password": {
                  "description": "Current password.",
                  "type": "string"
                },
                "type": {
                  "description": "Unit type.",
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucefully created.",
            "schema": {
              "$ref": "#/definitions/txHash"
            }
          },
          "400": {
            "description": "Invalid input",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "get": {
        "tags": [ "Unit types" ],
        "summary": "Hotel's unit types.",
        "description": "Retrieves hotel's unit types.",
        "parameters": [
          {
            "in": "path",
            "name": "address",
            "type": "string",
            "required": true,
            "description": "Hotel address."
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "array",
              "items" : {
                "type": "string"
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }

      }
    },
    "/hotels/{hotelAddress}/unitTypes/{type}": {
      "delete": {
        "tags": [ "Unit types" ],
        "summary": "remove unity type from hotel.",
        "description": "Remove a unit type from hotel.",
        "parameters": [
          {
            "in": "path",
            "name": "address",
            "type": "string",
            "required": true,
            "description": "Hotel address."
          },
          {
            "in": "path",
            "name": "type",
            "type": "string",
            "required": true,
            "description": "unit type."
          },
          {
            "in": "body",
            "name": "password",
            "required": true,
            "description": "Password.",
            "schema": {
              "required": [
                "password"
              ],
              "properties": {
                "password": {
                  "description": "Current password.",
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Sucefully removed.",
            "schema": {
              "$ref": "#/definitions/txHash"
            }
          },
          "400": {
            "description": "Invalid input",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "put": {
        "tags": [ "Unit types" ],
        "summary": "Edit unity type.",
        "description": "Edit information.",
        "parameters": [
          {
            "in": "path",
            "name": "address",
            "type": "string",
            "required": true,
            "description": "Hotel address."
          },
          {
            "in": "path",
            "name": "type",
            "type": "string",
            "required": true,
            "description": "Unit type."
          },
          {
            "in": "body",
            "name": "params",
            "required": true,
            "description": "Password.",
            "schema": {
              "required": [
                "password",
                "description",
                "minGuests",
                "maxGuests",
                "price"
              ],
              "properties": {
                "password": {
                  "description": "Current password.",
                  "type": "string"
                },
                "description": {
                  "description": "Unit type description.",
                  "type": "string"
                },
                "minGuests": {
                  "description": "Minimun guest allowed.",
                  "type": "number"
                },
                "maxGuests": {
                  "description": "Maximun guests allowed.",
                  "type": "number"
                },
                "price": {
                  "description": "Unit type price.",
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Sucefully removed.",
            "schema": {
              "$ref": "#/definitions/txHash"
            }
          },
          "400": {
            "description": "Invalid input",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/hotels/{hotelAddress}/unitTypes/{type}/images": {
      "post": {
        "tags": [ "Unit types" ],
        "summary": "Add image to unit type",
        "description": "Add an image for the unit type.",
        "parameters": [
          {
            "in": "path",
            "name": "address",
            "type": "string",
            "required": true,
            "description": "Hotel address."
          },
          {
            "in": "path",
            "name": "type",
            "type": "string",
            "required": true,
            "description": "Unit type."
          },
          {
            "name": "params",
            "in": "body",
            "required": true,
            "schema": {
              "required": [
                "password",
                "url"
              ],
              "properties": {
                "password": {
                  "description": "Current password.",
                  "type": "string"
                },
                "url": {
                  "description": "Picture's url.",
                  "type": "string"
                 }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucefully created.",
            "schema": {
              "$ref": "#/definitions/txHash"
            }
          },
          "400": {
            "description": "Invalid input",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/hotels/{hotelAddress}/unitTypes/{type}/images/{index}": {
      "delete": {
        "tags": [ "Unit types" ],
        "summary": "Remove image from unit type",
        "description": "Remove image from the unit type by index.",
        "parameters": [
          {
            "in": "path",
            "name": "address",
            "type": "string",
            "required": true,
            "description": "Hotel address."
          },
          {
            "in": "path",
            "name": "type",
            "type": "string",
            "required": true,
            "description": "Unit type."
          },
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "required": true,
            "description": "Image index."
          },
          {
            "name": "params",
            "in": "body",
            "required": true,
            "schema": {
              "required": [
                "password"
              ],
              "properties": {
                "password": {
                  "description": "Current password.",
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Sucefully removed.",
            "schema": {
              "$ref": "#/definitions/txHash"
            }
          },
          "400": {
            "description": "Invalid input",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/hotels/{hotelAddress}/unitTypes/{type}/amenities": {
      "post": {
        "tags": [ "Unit types" ],
        "summary": "Create amenities.",
        "description": "Add amenities for unit type.",
        "parameters": [
          {
            "in": "path",
            "name": "address",
            "type": "string",
            "required": true,
            "description": "Hotel address."
          },
          {
            "in": "path",
            "name": "type",
            "type": "string",
            "required": true,
            "description": "unit type."
          },
          {
            "in": "body",
            "name": "password",
            "required": true,
            "description": "Password.",
            "schema": {
              "required": [
                "password",
                "amenity"
              ],
              "properties": {
                "password": {
                  "description": "Current password.",
                  "type": "string"
                },
                "amenity": {
                  "description": "Code of amenity to add",
                  "type": "number"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucefully created.",
            "schema": {
              "$ref": "#/definitions/txHash"
            }
          },
          "400": {
            "description": "Invalid input",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/hotels/{hotelAddress}/unitTypes/{type}/amenities/{amenity}": {
      "delete": {
        "tags": [ "Unit types" ],
        "summary": "remove amenity.",
        "description": "Remove amenity for unit.",
        "parameters": [
          {
            "in": "path",
            "name": "address",
            "type": "string",
            "required": true,
            "description": "Hotel address."
          },
          {
            "in": "path",
            "name": "type",
            "type": "string",
            "required": true,
            "description": "unit type."
          },
          {
            "in": "path",
            "name": "amenity",
            "type": "number",
            "required": true,
            "description": "Amenity code."
          },
          {
            "in": "body",
            "name": "password",
            "required": true,
            "description": "Password.",
            "schema": {
              "required": [
                "password"
              ],
              "properties": {
                "password": {
                  "description": "Current password.",
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Sucefully removed.",
            "schema": {
              "$ref": "#/definitions/txHash"
            }
          },
          "400": {
            "description": "Invalid input",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "txHash": {
      "title": "Transaction hash",
      "description": "transaction hash for the operation.",
      "properties": {
        "txHash": {
          "type": "string",
          "description": "Transaction hash."
        }
      }
    },
    "Error": {
      "title": "Error",
      "description": "Default schema for errors returned by API.",
      "properties": {
        "code": {
          "type": "string",
          "description": "Computer-readable error. Use this for comparison."
        },
        "short": {
          "type": "string",
          "description": "Human-readable error with a short description of the error."
        },
        "long": {
          "type": "string",
          "description": "Verbose explanation of what happen."
        }
      }
    },
...
  }
}